pyproject.toml
[tool.poetry]
name = "tase"
version = "0.1.0"
description = "Telegram Audio Search Engine (lightning fast)"
homepage = "https://github.com/appheap/TASE"
repository = "https://github.com/appheap/TASE"
authors = [
    "Soran Ghaderi <soran.gdr.cs@gmail.com>",
    "Taleb Zarhesh <taleb.zarhesh@gmail.com>",
]
maintainers = [
    "Soran Ghaderi <soran.gdr.cs@gmail.com>",
    "Taleb Zarhesh <taleb.zarhesh@gmail.com>",
]
readme = "README.md"
license = "MIT"
include = [
    "LICENSE",
]
#packages = [
#    { include = "" }
#]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = [
    "python",
]

[tool.poetry.dependencies]
python = "^3.10"
emoji = "^2.0.0"
schedule = "^1.1.0"
elasticsearch = "^8.3.3"
Pyrogram = "^2.0.30"
TgCrypto = "^1.2.3"
tomli = "^2.0.1"
termcolor = "^1.1.0"
colorlog = "^6.6.0"
python-decouple = "^3.6"
python-arango = "^7.4.1"
kombu = "^5.2.4"
arrow = "^1.2.2"
pydantic = "^1.9.2"
Jinja2 = "^3.1.2"
APScheduler = "^3.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

